madds update 5/28

+++ do-not-delete-this/admin/peace_and_order/action_button/action_view.php
@@ -6,6 +6,8 @@ include '../../../includes/deactivated.inc.php';
 
 $id = $_GET['view_id'];
 
+$_SESSION['incident_id'] = $id;
+
 //select incident_table
 $sql = "SELECT * FROM incident_table
     WHERE incident_id = :id
@@ -123,25 +125,25 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
                         <tr>
                             <td>Case</td>
                             <td>
-                                <input type="text" name="case" value="<?php echo $list1['case_incident']; ?>">
+                                <input type="text" name="case" value="<?php echo $list1['case_incident']; ?>" readonly>
                             </td>
                         </tr>
                         <tr>
                             <td>Incident title</td>
                             <td>
-                                <input type="text" name="case" value="<?php echo $list1['incident_title']; ?>">
+                                <input type="text" name="case" value="<?php echo $list1['incident_title']; ?>" readonly>
                             </td>
                         </tr>
                         <tr>
                             <td>Date of Incident</td>
                             <td>
-                                <input type="date" name="case" value="<?php echo $list1['date_incident']; ?>">
+                                <input type="date" name="case" value="<?php echo $list1['date_incident']; ?>" readonly>
                             </td>
                         </tr>
                         <tr>
                             <td>Time of Incident</td>
                             <td>
-                                <input type="time" name="case" value="<?php echo $list1['time_incident']; ?>">
+                                <input type="time" name="case" value="<?php echo $list1['time_incident']; ?>" readonly>
                             </td>
                         </tr>
                     </table>
@@ -168,10 +170,10 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
                                         <td><?php echo "Offender"; ?></td>
                                         <td>
                                             <div class="action_btn">
-                                                <button type="button" class="text-blue-700 hover:text-white border border-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2 dark:border-blue-500 dark:text-blue-500 dark:hover:text-white dark:hover:bg-blue-500 dark:focus:ring-blue-800">
-                                                    <a href="edit_person.php?up_off_id=<?php echo $list['offender_id']; ?>">
+                                                <button type="button" class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
+                                                    <a href="edit_personOff.php?up_off_id=<?php echo $list['offender_id']; ?>">
                                                         Edit details</a></button>
-                                                <button onclick=" delete_person()" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
+                                                <button onclick="return confirm('are you sure?')" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
                                                     <a href="delete_person.php? view_id=<?php echo $id; ?>&del_off_id=<?php echo $list['offender_id']; ?>">
                                                         Remove</a></button>
                                             </div>
@@ -188,10 +190,10 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
                                         <td>
                                             <div class="action_btn">
-                                                <button class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"><a href="edit_person.php?up_off_id=<?php echo $list['offender_id']; ?>">
+                                                <button class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"><a href="edit_personOff.php?up_off_id=<?php echo $list['offender_id']; ?>">
                                                         Edit details</a></button>
-                                                <button onclick="delete_person()" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
-                                                    Remove</a></button>
+                                                <button onclick="return confirm('are you sure?')" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
+                                                    <a href="delete_person.php? view_id=<?php echo $id; ?>&del_off_id=<?php echo $list['offender_id']; ?>">Remove</a></button>
                                             </div>
                                         </td>
                                     </tr>
@@ -205,7 +207,7 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
                                 if ($list['complainant_type'] == "resident") {
                             ?>
                                     <tr>
-                                        <td><?php echo $list['non_firstname'] . " " . $list['lastname']; ?></td>
+                                        <td><?php echo $list['firstname'] . " " . $list['lastname']; ?></td>
                                         <td><?php echo $list['sex']; ?></td>
                                         <td><?php echo $list['address']; ?></td>
                                         <td><?php echo "Complainant"; ?></td>
@@ -213,9 +215,9 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
                                         <td>
                                             <div class="action_btn">
                                                 <button type="button" class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
-                                                    <a href="edit_person.php?up_comp_id=<?php echo $list['complainant_id']; ?>">
+                                                    <a href="edit_personComp.php?up_comp_id=<?php echo $list['complainant_id']; ?>">
                                                         Edit details</a></button>
-                                                <button onclick="delete_person()" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
+                                                <button onclick="return confirm('are you sure?')" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
                                                     <a href="delete_person.php? view_id=<?php echo $id; ?>&del_comp_id=<?php echo $list['complainant_id']; ?>">
                                                         Remove</a></button>
                                             </div>
@@ -231,9 +233,10 @@ $list3 = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
                                         <td>
                                             <div class="action_btn">
-                                                <button class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"><a href="edit_person.php?up_comp_id=<?php echo $list['complainant_id']; ?>">
+                                                <button class="px-3 py-2 text-xs font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"><a href="edit_personComp.php?up_comp_id=<?php echo $list['complainant_id']; ?>">
                                                         Edit details</a></button>
-                                                <button onclick="return confirm('are you sure?')" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900"> <a href="delete_person.php? view_id=<?php echo $id; ?>&del_comp_id=<?php echo $list['complainant_id']; ?>">
+                                                <button onclick="return confirm('are you sure?')" class="text-red-700 hover:text-white border border-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 font-medium rounded-lg text-sm px-3 py-1 text-center mr-2 mb-2 dark:border-red-500 dark:text-red-500 dark:hover:text-white dark:hover:bg-red-600 dark:focus:ring-red-900">
+                                                    <a href="delete_person.php? view_id=<?php echo $id; ?>&del_comp_id=<?php echo $list['complainant_id']; ?>">
                                                         Remove</a></button>
                                             </div>
                                         </td>
